+++ Gemfile
@@ -2,7 +2,7 @@
 
 source "https://rubygems.org"
 
+gem "private_address_check", "0.4.1"
 # Specify your gem's dependencies in pronto-bundler_audit.gemspec
 gemspec
 
-gem "private_address_check", "0.4.1"
