+++ lib/pronto/bundler_audit.rb
@@ -10,11 +10,6 @@ module Pronto
   # 2. Runs bundle-audit to scan the Gemfile.lock, and then
   # 4. Returns an Array of Pronto::Message objects if any issues or advisories
   # are found.
-
-  Line = Struct.new(:line_number) do
-    alias :new_lineno :line_number
-  end
-
   class BundlerAudit < Runner
     GEMFILE_LOCK_FILENAME = "Gemfile.lock"
 
@@ -22,19 +17,33 @@ module Pronto
     def run
       results = Auditor.new.call
 
-      return [] unless results && results.size > 0
+      return results unless results.size > 0
 
-      #require 'pry'
-      #binding.pry
       results.map do |result|
         Message.new(GEMFILE_LOCK_FILENAME,
-                    Line.new(line_number: result.line),
+                    DeepLine.new(result.line, @patches.repo.path),
                     result.level,
                     result.message,
                     @patches.commit,
                     self.class)
       end
     end
+  end
+
+  # This piece of ugliness is here due to prontos message handling, its a bit
+  # of a mess in there and rather than deal with wrapping it, we create a class
+  # to supply it with what it wants i.e. line_number and path.
+  class DeepLine
+    attr_reader :line_number, :path
+
+    def initialize(line_number, path)
+      @path = path
+      @line_number = line_number
+    end
+
+    alias :repo :itself
+    alias :patch :itself
+    alias :new_lineno :line_number
   end
 end
