+++ lib/pronto/bundler_audit/auditor.rb
@@ -1,17 +1,34 @@
 # frozen_string_literal: true
 
-require "pronto/bundler_audit/scanner"
+require_relative "results"
 
 module Pronto
   class BundlerAudit
     # Pronto::BundlerAudit::Auditor:
     # 1. updates the local ruby security database, and then
-    # 2. runs {Pronto::BundlerAudit::Scanner#call}.
+    # 2. runs {Pronto::Bundler::Audit::Scanner}.
+    # 3. returns results converted to the appropriate format
     class Auditor
-      # @return (see: #run_scan)
       def call
         Bundler::Audit::Database.update!(quiet: true)
-        Scanner.new.call
+        scan.map { |result| match_result(result).call } || []
+      end
+
+      private
+
+      def scan
+        Bundler::Audit::Scanner.new.scan
+      end
+
+      def match_result(scan_result)
+        case scan_result
+        when Bundler::Audit::Scanner::InsecureSource
+          Results::InsecureSource.new(scan_result)
+        when Bundler::Audit::Scanner::UnpatchedGem
+          Results::UnpatchedGem.new(scan_result)
+        else
+          raise ArgumentError, "Unexpected type: #{scan_result.class}"
+        end
       end
     end
   end
